refactor(api): extract InventoryDownloadService.queryInventoryData from download route
So it can be reused
